Extract Modal component from CheckIn and hook it up to the ModalContext
