+++ client/src/components/AddBandModal/index.js
@@ -1,0 +1,2 @@
+import AddBandModal from './AddBandModal';
+export default AddBandModal;
